+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/protocol/blob_rest_client.hpp
@@ -10270,9 +10270,6 @@ namespace Azure { namespace Storage { namespace Blobs {
           Azure::Nullable<int32_t> Timeout;
           int64_t BlobSize = -1;
           Azure::Nullable<std::string> LeaseId;
-          Azure::Nullable<int64_t> IfSequenceNumberLessThanOrEqualTo;
-          Azure::Nullable<int64_t> IfSequenceNumberLessThan;
-          Azure::Nullable<int64_t> IfSequenceNumberEqualTo;
           Azure::Nullable<Azure::DateTime> IfModifiedSince;
           Azure::Nullable<Azure::DateTime> IfUnmodifiedSince;
           Azure::ETag IfMatch;
@@ -10300,24 +10297,6 @@ namespace Azure { namespace Storage { namespace Blobs {
           if (options.LeaseId.HasValue())
           {
             request.SetHeader("x-ms-lease-id", options.LeaseId.Value());
-          }
-          if (options.IfSequenceNumberLessThanOrEqualTo.HasValue())
-          {
-            request.SetHeader(
-                "x-ms-if-sequence-number-le",
-                std::to_string(options.IfSequenceNumberLessThanOrEqualTo.Value()));
-          }
-          if (options.IfSequenceNumberLessThan.HasValue())
-          {
-            request.SetHeader(
-                "x-ms-if-sequence-number-lt",
-                std::to_string(options.IfSequenceNumberLessThan.Value()));
-          }
-          if (options.IfSequenceNumberEqualTo.HasValue())
-          {
-            request.SetHeader(
-                "x-ms-if-sequence-number-eq",
-                std::to_string(options.IfSequenceNumberEqualTo.Value()));
           }
           if (options.IfModifiedSince.HasValue())
           {
@@ -10368,9 +10347,6 @@ namespace Azure { namespace Storage { namespace Blobs {
           SequenceNumberAction Action;
           Azure::Nullable<int64_t> SequenceNumber;
           Azure::Nullable<std::string> LeaseId;
-          Azure::Nullable<int64_t> IfSequenceNumberLessThanOrEqualTo;
-          Azure::Nullable<int64_t> IfSequenceNumberLessThan;
-          Azure::Nullable<int64_t> IfSequenceNumberEqualTo;
           Azure::Nullable<Azure::DateTime> IfModifiedSince;
           Azure::Nullable<Azure::DateTime> IfUnmodifiedSince;
           Azure::ETag IfMatch;
@@ -10397,24 +10373,6 @@ namespace Azure { namespace Storage { namespace Blobs {
           if (options.LeaseId.HasValue())
           {
             request.SetHeader("x-ms-lease-id", options.LeaseId.Value());
-          }
-          if (options.IfSequenceNumberLessThanOrEqualTo.HasValue())
-          {
-            request.SetHeader(
-                "x-ms-if-sequence-number-le",
-                std::to_string(options.IfSequenceNumberLessThanOrEqualTo.Value()));
-          }
-          if (options.IfSequenceNumberLessThan.HasValue())
-          {
-            request.SetHeader(
-                "x-ms-if-sequence-number-lt",
-                std::to_string(options.IfSequenceNumberLessThan.Value()));
-          }
-          if (options.IfSequenceNumberEqualTo.HasValue())
-          {
-            request.SetHeader(
-                "x-ms-if-sequence-number-eq",
-                std::to_string(options.IfSequenceNumberEqualTo.Value()));
           }
           if (options.IfModifiedSince.HasValue())
           {