+++ sdk/storage/azure-storage-files-shares/test/share_directory_client_test.cpp
@@ -14,7 +14,7 @@ namespace Azure { namespace Storage { namespace Test {
 
   void FileShareDirectoryClientTest::SetUpTestSuite()
   {
-    m_directoryName = LowercaseRandomString();
+    m_directoryName = RandomString();
     m_shareName = LowercaseRandomString();
     m_shareClient = std::make_shared<Files::Shares::ShareClient>(
         Files::Shares::ShareClient::CreateFromConnectionString(
@@ -77,7 +77,7 @@ namespace Azure { namespace Storage { namespace Test {
       std::vector<Files::Shares::ShareDirectoryClient> directoryClients;
       for (int32_t i = 0; i < 5; ++i)
       {
-        auto name = RandomString(10);
+        auto name = RandomString();
         Files::Shares::ShareDirectoryClient client
             = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(name);
         EXPECT_NO_THROW(client.Create());
@@ -93,7 +93,7 @@ namespace Azure { namespace Storage { namespace Test {
       // Create directory that already exist throws.
       for (int32_t i = 0; i < 5; ++i)
       {
-        auto name = RandomString(10);
+        auto name = RandomString();
         Files::Shares::ShareDirectoryClient client
             = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(name);
         EXPECT_NO_THROW(client.Create());
@@ -103,23 +103,20 @@ namespace Azure { namespace Storage { namespace Test {
     {
       // CreateIfNotExists & DeleteIfExists.
       {
-        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(
-            LowercaseRandomString());
+        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
         EXPECT_NO_THROW(client.Create());
         EXPECT_NO_THROW(client.CreateIfNotExists());
         EXPECT_NO_THROW(client.Delete());
         EXPECT_NO_THROW(client.DeleteIfExists());
       }
       {
-        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(
-            LowercaseRandomString());
+        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
         EXPECT_NO_THROW(client.CreateIfNotExists());
         EXPECT_THROW(client.Create(), StorageException);
         EXPECT_NO_THROW(client.DeleteIfExists());
       }
       {
-        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(
-            LowercaseRandomString());
+        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
         auto created = client.Create().Value.Created;
         EXPECT_TRUE(created);
         auto createResult = client.CreateIfNotExists();
@@ -130,23 +127,21 @@ namespace Azure { namespace Storage { namespace Test {
         EXPECT_TRUE(deleted);
       }
       {
-        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(
-            LowercaseRandomString());
+        auto client = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
         auto deleteResult = client.DeleteIfExists();
         EXPECT_FALSE(deleteResult.Value.Deleted);
       }
       {
         auto shareClient = Files::Shares::ShareClient::CreateFromConnectionString(
             StandardStorageConnectionString(), LowercaseRandomString());
-        auto client
-            = shareClient.GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+        auto client = shareClient.GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
         auto deleteResult = client.DeleteIfExists();
         EXPECT_FALSE(deleteResult.Value.Deleted);
       }
       {
         auto client = m_shareClient->GetRootDirectoryClient()
-                          .GetSubdirectoryClient(LowercaseRandomString())
-                          .GetSubdirectoryClient(LowercaseRandomString());
+                          .GetSubdirectoryClient(RandomString())
+                          .GetSubdirectoryClient(RandomString());
         auto deleteResult = client.DeleteIfExists();
         EXPECT_FALSE(deleteResult.Value.Deleted);
       }
@@ -169,10 +164,8 @@ namespace Azure { namespace Storage { namespace Test {
 
     {
       // Create directory with metadata works
-      auto client1
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
-      auto client2
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client1 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
+      auto client2 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
       Files::Shares::CreateDirectoryOptions options1;
       Files::Shares::CreateDirectoryOptions options2;
       options1.Metadata = metadata1;
@@ -195,10 +188,8 @@ namespace Azure { namespace Storage { namespace Test {
 
     {
       // Create directory with permission/permission key works
-      auto client1
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
-      auto client2
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client1 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
+      auto client2 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
       Files::Shares::CreateDirectoryOptions options1;
       Files::Shares::CreateDirectoryOptions options2;
       options1.DirectoryPermission = permission;
@@ -210,8 +201,7 @@ namespace Azure { namespace Storage { namespace Test {
       auto result2 = client2.GetProperties().Value.SmbProperties.PermissionKey;
       EXPECT_EQ(result1.Value(), result2.Value());
 
-      auto client3
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client3 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
       Files::Shares::CreateDirectoryOptions options3;
       options3.SmbProperties.PermissionKey = result1;
       EXPECT_NO_THROW(client3.Create(options3));
@@ -227,10 +217,8 @@ namespace Azure { namespace Storage { namespace Test {
       properties.CreatedOn = std::chrono::system_clock::now();
       properties.LastWrittenOn = std::chrono::system_clock::now();
       properties.PermissionKey = "";
-      auto client1
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
-      auto client2
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client1 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
+      auto client2 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
 
       EXPECT_NO_THROW(client1.Create());
       EXPECT_NO_THROW(client2.Create());
@@ -244,8 +232,7 @@ namespace Azure { namespace Storage { namespace Test {
       auto result2 = client2.GetProperties().Value.SmbProperties.PermissionKey;
       EXPECT_EQ(result1.Value(), result2.Value());
 
-      auto client3
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client3 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
       Files::Shares::CreateDirectoryOptions options3;
       options3.SmbProperties.PermissionKey = result1;
       std::string permissionKey;
@@ -267,10 +254,8 @@ namespace Azure { namespace Storage { namespace Test {
         = m_fileShareDirectoryClient->GetProperties().Value.SmbProperties.PermissionKey;
     {
       // Create directory with SmbProperties works
-      auto client1
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
-      auto client2
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client1 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
+      auto client2 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
       Files::Shares::CreateDirectoryOptions options1;
       Files::Shares::CreateDirectoryOptions options2;
       options1.SmbProperties = properties;
@@ -293,10 +278,8 @@ namespace Azure { namespace Storage { namespace Test {
 
     {
       // SetProperties works
-      auto client1
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
-      auto client2
-          = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(LowercaseRandomString());
+      auto client1 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
+      auto client2 = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
 
       EXPECT_NO_THROW(client1.Create());
       EXPECT_NO_THROW(client2.Create());
@@ -316,11 +299,35 @@ namespace Azure { namespace Storage { namespace Test {
     }
   }
 
+  TEST_F(FileShareDirectoryClientTest, SmbPropertiesDefaultValue)
+  {
+    auto directoryClient
+        = m_shareClient->GetRootDirectoryClient().GetSubdirectoryClient(RandomString());
+    directoryClient.Create();
+    auto smbProperties = directoryClient.GetProperties().Value.SmbProperties;
+    EXPECT_EQ(smbProperties.Attributes, Files::Shares::Models::FileAttributes::Directory);
+    ASSERT_TRUE(smbProperties.CreatedOn.HasValue());
+    EXPECT_TRUE(IsValidTime(smbProperties.CreatedOn.Value()));
+    ASSERT_TRUE(smbProperties.LastWrittenOn.HasValue());
+    EXPECT_TRUE(IsValidTime(smbProperties.LastWrittenOn.Value()));
+    ASSERT_TRUE(smbProperties.ChangedOn.HasValue());
+    EXPECT_TRUE(IsValidTime(smbProperties.ChangedOn.Value()));
+
+    directoryClient.SetProperties(Files::Shares::Models::FileSmbProperties());
+
+    auto smbProperties2 = directoryClient.GetProperties().Value.SmbProperties;
+    EXPECT_EQ(smbProperties2.PermissionKey.Value(), smbProperties.PermissionKey.Value());
+    EXPECT_EQ(smbProperties2.Attributes, smbProperties.Attributes);
+    EXPECT_EQ(smbProperties2.CreatedOn.Value(), smbProperties.CreatedOn.Value());
+    EXPECT_EQ(smbProperties2.LastWrittenOn.Value(), smbProperties.LastWrittenOn.Value());
+    EXPECT_NE(smbProperties2.ChangedOn.Value(), smbProperties.ChangedOn.Value());
+  }
+
   TEST_F(FileShareDirectoryClientTest, ListFilesAndDirectoriesSinglePageTest)
   {
     // Setup
-    auto directoryNameA = LowercaseRandomString();
-    auto directoryNameB = LowercaseRandomString();
+    auto directoryNameA = RandomString();
+    auto directoryNameB = RandomString();
     std::vector<std::string> directoryNameSetA;
     std::vector<std::string> directoryNameSetB;
     std::vector<std::string> fileNameSetA;
@@ -332,16 +339,16 @@ namespace Azure { namespace Storage { namespace Test {
     for (size_t i = 0; i < 5; ++i)
     {
       {
-        auto directoryName = LowercaseRandomString();
-        auto fileName = LowercaseRandomString();
+        auto directoryName = RandomString();
+        auto fileName = RandomString();
         EXPECT_NO_THROW(clientA.GetSubdirectoryClient(directoryName).Create());
         EXPECT_NO_THROW(clientA.GetFileClient(fileName).Create(1024));
         directoryNameSetA.emplace_back(std::move(directoryName));
         fileNameSetA.emplace_back(std::move(fileName));
       }
       {
-        auto directoryName = LowercaseRandomString();
-        auto fileName = LowercaseRandomString();
+        auto directoryName = RandomString();
+        auto fileName = RandomString();
         EXPECT_NO_THROW(clientB.GetSubdirectoryClient(directoryName).Create());
         EXPECT_NO_THROW(clientB.GetFileClient(fileName).Create(1024));
         directoryNameSetB.emplace_back(std::move(directoryName));